Code change to allow fast marching & fast sweeping on 2D unstructured meshes with slowness defined at nodes (Grid2Dui)

+++ ttcr/Node2Di.h
@@ -32,8 +32,7 @@ public:
 		}
 	}
 	
-    
-    Node2Di(const T1 t, const sxz<T1>& s, const int nt, const size_t i) :
+    Node2Di(const T1 t, const sxz<T1>& s, const size_t nt, const size_t i) :
 	nThreads(nt),
     tt(0),
 	x(s.x), z(s.z), slowness(0.0),
